+++ src/TerraformCommand.groovy
@@ -1,0 +1,3 @@
+interface TerraformCommand {
+    public String getEnvironment()
+}